pyproject.toml
[project]
name = "multi-agent-vlm-orchestrator"
version = "0.1.0"
description = "Agentic orchestration system for Hugging Face vision-language model experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.8.2",
    "typer>=0.12.3",
]

[project.optional-dependencies]
local = [
    "accelerate>=0.33.0",
    "pillow>=10.4.0",
    "torch>=2.4.0",
    "transformers>=4.44.0",
]
dev = [
    "pytest>=8.3.2",
    "ruff>=0.6.2",
]
yaml = [
    "pyyaml>=6.0.2",
]

[project.scripts]
multi-agent-vlm-orchestrator = "multi_agent_vlm_orchestrator.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_agent_vlm_orchestrator"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
